rag/requirements.txt
markitdown[all]==0.1.1
openai==1.77.0
python-dotenv==1.1.0
tiktoken==0.9.0
qdrant_client==1.14.2
rich==14.0.0
google-generativeai